edit files

+++ downloader/dl_ehentai.py
@@ -48,21 +48,23 @@ async def main(gallery_link, loop):
 
     dirLoc = '[e-hentai] ' + GetFileName(f'{gTitle}')
 
-    MakeDirectory(f'./{download_folder}/{dirLoc}/')
-
-    imageLoc = []
-    tasks = []
-    for idx, imgurl in enumerate(imgsURL):
-        imgName = f'./{download_folder}/{dirLoc}/{idx}.jpg'
-        tasks.append(asyncio.ensure_future(FileDownload(filename=imgName, fileurl=imgurl)))
-        imageLoc.append(imgName)
+    if isfile(f'./{download_folder}/{dirLoc}.pdf') != True:
         
-    
-    await asyncio.gather(*tasks)
+        MakeDirectory(f'./{download_folder}/{dirLoc}/')
 
-    await asyncio.gather(asyncio.ensure_future(MakePDF(ImageList=imageLoc, Filename=f'./{download_folder}/{dirLoc}.pdf')))
+        imageLoc = []
+        tasks = []
+        for idx, imgurl in enumerate(imgsURL):
+            imgName = f'./{download_folder}/{dirLoc}/{idx}.jpg'
+            tasks.append(asyncio.ensure_future(FileDownload(filename=imgName, fileurl=imgurl)))
+            imageLoc.append(imgName)
+            
+        
+        await asyncio.gather(*tasks)
 
-    rmtree(f'./{download_folder}/{dirLoc}/', ignore_errors=True)
+        await asyncio.gather(asyncio.ensure_future(MakePDF(ImageList=imageLoc, Filename=f'./{download_folder}/{dirLoc}.pdf')))
+
+        rmtree(f'./{download_folder}/{dirLoc}/', ignore_errors=True)
